[Refactor][Core/Dashboard] Merge HealthzHead into APIHead
