+++ python/ray/dashboard/modules/snapshot/healthz_agent.py
@@ -3,7 +3,7 @@ from aiohttp.web import Request, Response
 import ray.dashboard.optional_utils as optional_utils
 import ray.dashboard.utils as dashboard_utils
 import ray.exceptions
-from ray.dashboard.modules.healthz.utils import HealthChecker
+from ray.dashboard.modules.snapshot.utils import HealthChecker
 
 routes = optional_utils.DashboardAgentRouteTable
